pyproject.toml
[project]
name = "datagen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.0.0",
    "opencv-python>=4.11.0.86",
    "pillow>=10.0.0",
    "trdg @ git+https://github.com/Belval/TextRecognitionDataGenerator.git@master",
]
